feat(hydration): move hydration utilities to core, keeping old exports
