+++ src/hydration/index.ts
@@ -1,4 +1,4 @@
-export { dehydrate, hydrate } from './hydration'
+export { dehydrate, hydrate } from '../core/hydration'
 export { useHydrate, Hydrate } from './react'
 
 // Types
@@ -6,6 +6,7 @@ export type {
   DehydrateOptions,
   DehydratedState,
   HydrateOptions,
+  ShouldDehydrateMutationFunction,
   ShouldDehydrateQueryFunction,
-} from './hydration'
+} from '../core/hydration'
 export type { HydrateProps } from './react'
