+++ src/hydration/react.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { useQueryClient } from 'react-query'
-import { hydrate, HydrateOptions } from './hydration'
+import { hydrate, HydrateOptions } from '../core/hydration'
 
 export function useHydrate(state: unknown, options?: HydrateOptions) {
   const queryClient = useQueryClient()
